v0.0.29
add types and resolvers for product

+++ server/graphql/types/product-batch/product-batch-list.ts
@@ -3,6 +3,6 @@ import { gql } from 'apollo-server-koa'
 export const ProductBatchList = gql`
   type ProductBatchList {
     items: [ProductBatch]
-    total: Int
+    total: Float
   }
 `
